+++ application/views/client/cart.php
@@ -16,21 +16,23 @@
 								<tr class="tbl-ttl">
 									<td class="td-1">&nbsp;</td><td class="td-2">Name</td><td class="td-3">Art.</td><td class="td-4">Per Unit</td><td class="td-5">Qty</td><td class="td-6">Total</td>
 								</tr>
-								<?foreach($cart as $item_id => $item):?>
-									<tr id="<?=$item_id?>" class="tbl-item">
-										<td class="td-1"><a href=""><img src="images/ronin/ronin1.jpg" alt=""/></a></td>
-										<td class="td-2"><?=$item['title']?></td>
-										<td class="td-3"><span class="">SR01</span></td>
-										<td class="td-4"><span class="unit-pr"><?=$item['price']?></span>&euro;</td>
-										<td class="td-5"><input type="text" name="qty_<?=$item_id?>" id="qty_<?=$item_id?>" value="<?=$item['qty']?>" onchange="update_cart('<?=$item_id?>', this.value);"/></td>
-										<td class="td-6">
-											<div class="td-6">
-												<span id="item_total_<?=$item_id?>"><?=$item['item_total']?></span>&euro;
-												<span class="close" onclick="delete_item('<?=$item_id?>');">&nbsp;</span>
-											</div>	
-										</td>
-									</tr>
-								<?endforeach;?>
+								<?if(isset($cart)):?>
+									<?foreach($cart as $item_id => $item):?>
+										<tr class="tbl-item <?=$item_id?>">
+											<td class="td-1"><a href=""><img src="<?=$item['img']->url?>" alt=""/></a></td>
+											<td class="td-2"><?=$item['title']?></td>
+											<td class="td-3"><span class="">SR01</span></td>
+											<td class="td-4"><span class="unit-pr"><?=$item['price']?></span>&euro;</td>
+											<td class="td-5"><input type="text" name="qty_<?=$item_id?>" id="qty_<?=$item_id?>" value="<?=$item['qty']?>" onchange="update_cart('<?=$item_id?>', this.value);"/></td>
+											<td class="td-6">
+												<div class="td-6">
+													<span id="total_<?=$item_id?>"><?=$item['item_total']?></span> &euro;
+													<span class="close" onclick="delete_item('<?=$item_id?>');">&nbsp;</span>
+												</div>	
+											</td>
+										</tr>
+									<?endforeach;?>
+								<?endif;?>
 							</table>
 						</div>
 						<div id="overall" class="clearfix">
@@ -38,48 +40,28 @@
 								<table>
 									<tr>
 										<td class="tt">In Total:</td>
-										<td><span class="total_price"><?=$total_price?></span>&euro;</td>
+										<td><span class="total_price"><?=$total_price?></span> &euro;</td>
 									</tr>
 									<tr>
 										<td>Shipping:</td>
-										<td><span class="shipping-price">6</span>&euro;</td>
+										<td><span class="shipping"></span> &euro;</td>
 									</tr>
-								</table>
+								</table>							
 							</div>
-			  <div class="total-right">
-			    <div class="o-w-s">
-                   Summary:
+							<div class="total-right">
+								<div class="o-w-s">Summary:</div>
+								<div class="o-w-s-2"><span class="total">0</span> &euro;</div>
+							</div>
+						</div>
+						<div class="proceed">
+							<a  href="<?=base_url()?>pages/cart/2" id="proceed_link" type="button" name="" style="display:none"></a>
+							<span class="freak_link" onclick=proceed_link.click()>PROCEED</span>
+						</div>
+					</div>
 				</div>
-				<div class="o-w-s-2">
-				  <span id="total_price"><?=$total_price?></span>&euro;
-				</div>
-			  </div>
+				<?require_once 'include/right_col.php'?>
 			</div>
-			  <div class="proceed">
-			    <input id="proceed_link" type="submit" name="" style="display:none">
-				<span class="freak_link" onclick=proceed_link.click()>PROCEED</span>
-			  </div>
-		  </div>
+			<?require_once 'include/footer.php'?>
 		</div>
-		<?require_once 'include/right_col.php'?>
-
-	  </div>
-		<?require_once 'include/footer.php'?>
-    </div>	
-
-<script>
-$('ul .up').click(function() {
-$(this).next().slideToggle().toggleClass('noactive');
-$(this).toggleClass('up');
-$(this).toggleClass('down');
-});
-
-$('ul .down').click(function() {
-$(this).next().slideToggle().toggleClass('noactive');
-$(this).toggleClass('down');
-$(this).toggleClass('up');
-});
-</script>
-
-  </body>
+	</body>
 </html>
